Initial draft of mops2tfr.py This program transforms a directory tree containing MOPS "quad" images and parameters (in) with accompanying text tag files (out) indicating whether the image contains a NEO or not. This data is transformed into a TensorFlow training set in TFRecord format and splits the training examples into train, validate, and (held out) test sets.
Each training image may contain multiple object candidates. For each, a 32x32 cutout centered on the object is cut out
and stored with the label1 (actual object) or 0 (false detection). An additional parameter vector for each candidate is
saved as a float vector. Each training example then consists of a 32x32 float image, a 64 float vector of parameters and
a single byte label that is binary 1 or 0.